noCopyState as default and changes for additional luigi task
