+++ workflows/CopyState.py
@@ -28,7 +28,7 @@ with open(configFile, 'r') as getConfiguration:
     configuration = json.load(getConfiguration)
 
 if "noCopyState" in configuration and configuration["noCopyState"]:
-    log.warning("noCopySate flag set")
+    log.warning("noCopyState flag set")
     sys.exit()
     
 targetStatePath = os.path.join("/output", "state", configuration["inputFolder"], datetime.now().strftime("%Y%m%d_%H%M%S"))
