+++ workflows/requirements.txt
@@ -1,1 +1,7 @@
-luigi==3.0.3
+lockfile==0.12.2
+luigi==3.6.0
+python-daemon==3.1.2
+python-dateutil==2.9.0.post0
+six==1.17.0
+tenacity==8.5.0
+tornado==6.5.2
